Update and rename assignment-1 code to assignment-1-code
I think this is what you are asking for as far as the name on the file go. A little more detail in what is needed would have been helpful. As in please rename this file with no spaces and save it as an html file. that way it opens correctly. This is just an example so that I know what exactly you are looking for in the future. I do not know what an extension is either.